refactor(ui): break down renderer into feature-specific modules
Decouples the main renderer by breaking it into smaller, more focused modules that are responsible for a single feature's UI. This improves separation of concerns and makes the components more self-contained.

- Creates `features/transactions/transaction-renderer.js` to handle IOU/UOM lists.
- Creates `features/persons/person-renderer.js` to handle the people list.
- Creates `features/stats/stats-renderer.js` to handle the statistics page and chart.
- The main `ui/renderer.js` now acts as a router, delegating to the appropriate submodule based on the current view.
